move linker stuff from linkage/ to linker/
