+++ mismo/linkage/__init__.py
@@ -6,20 +6,9 @@ from mismo.linkage._combine import (
     unify_links_min_intersection as unify_links_min_intersection,
 )
 from mismo.linkage._combine import union as union
-from mismo.linkage._join_linker import JoinLinkage as JoinLinkage
-from mismo.linkage._join_linker import JoinLinker as JoinLinker
-from mismo.linkage._key_linker import KeyLinkage as KeyLinkage
-from mismo.linkage._key_linker import KeyLinker as KeyLinker
-from mismo.linkage._label_linker import LabelLinker as LabelLinker
 from mismo.linkage._linkage import BaseLinkage as BaseLinkage
 from mismo.linkage._linkage import Linkage as Linkage
 from mismo.linkage._linkage import LinkTableLinkage as LinkTableLinkage
 from mismo.linkage._linkage import filter_links as filter_links
-from mismo.linkage._linker import EmptyLinkage as EmptyLinkage
-from mismo.linkage._linker import EmptyLinker as EmptyLinker
-from mismo.linkage._linker import FullLinkage as FullLinkage
-from mismo.linkage._linker import FullLinker as FullLinker
-from mismo.linkage._linker import Linker as Linker
-from mismo.linkage._linker import UnnestLinker as UnnestLinker
 from mismo.linkage._sample import sample_all_links as sample_all_links
 from mismo.linkage._upset_block import upset_chart as upset_chart
